+++ CHANGELOG.md
@@ -26,6 +26,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Typo in Message.selectors.json file
 
+### Changed
+- Move ProductItem to atomic design, rename to ProductListItem
+
 ## [0.4.0] - 04-06-2019
 ### Added
 - This CHANGELOG